Feature/coursepage (#25)
* Add files via upload

This doesn't fully work, there is a bug with the hooking on line 16

* Add files via upload

The buttons are now working as intended, with a button pulling up a window to show the available tutors, which needs to be programmed in. I added some comments in to help show what goes where.

* Add files via upload

Change the name to App.js or change your index.js to work with this, but this is more of an alternate version that I think will make working with the REST API a lot easier, so we should probably move forward with this.

* Add files via upload

This doesn't fully work, there is a bug with the hooking on line 16

* Add files via upload

The buttons are now working as intended, with a button pulling up a window to show the available tutors, which needs to be programmed in. I added some comments in to help show what goes where.

* Add files via upload

Change the name to App.js or change your index.js to work with this, but this is more of an alternate version that I think will make working with the REST API a lot easier, so we should probably move forward with this.

* feat- buttons truncated to single stack

* Add files via upload

feat- importing names from backend

* Added searchbar functionality, updated course buttons

* Added searchbar functionality, updated course buttons, hopefully done correctly

* Added searchbar functionality, updated course buttons, hopefully done correctly for real this time

* Added the backend folder back