Add tutors and tutoring relation
The Tutoring table is done manually since we will likely be adding
additional columns (ex. price) later on. Furthermore, getting the tutor
or course from the relation is messy due to preloading; using a join is
likely a better solution and may allow the tutor fields to be extracted
directly.