Add backend testing support and implement tests (#45)
* Add backend testing support and implement example tests

This uses testify.suite for test setup to initialize the router. The
InitDatabase function now takes a dsn string to allow tests to run with
an in-memory database. Furthermore, testify does not appear to support
setup before subtests so tests are wrapped with a manualSetupTest
function.

* Add backend testing for individual course/tutor routes

This fixes two implementation issues; order was not being applied
(mainly due to earlier issues with Gorm) and the result slice was not
initialized thus returning empty (Gorm only assigns if results are
present).