Update authentication handling for JWTs & session cookies
This simplifies the authentication workflow to immediately create the
session cookie after signup, which removes the need for the /user route.
The user/:username route has also been replaced by /profile, which uses
the authenticated user to prevent public access if private data needs
to be included. Tests have been updated/added as well.

The frontend was patched for this, but ultimately needs to be refactored
to handle this flow properly.