Clean up backend API routes (#86)
* Clean up API for static course/tutor routes

This moves the user profile and availability into the tutor object as
tutor.user and tutor.availability respectively. Availability is handled
with a quick patch since it will eventually be replaced with a system
that can support times.

* Update authentication handling for JWTs & session cookies

This simplifies the authentication workflow to immediately create the
session cookie after signup, which removes the need for the /user route.
The user/:username route has also been replaced by /profile, which uses
the authenticated user to prevent public access if private data needs
to be included. Tests have been updated/added as well.

The frontend was patched for this, but ultimately needs to be refactored
to handle this flow properly.

* Patch frontend cypress tests

Removes tests for /tutors pages (removed), fixes /tutors/:username, and
disables the cookie test (state dependent on users and doesn't use
https for secure cookies). This will need to be fixed.