Add actions for automated testing (#87)
* Add action for backend Go tests

* Add action for frontend Jest tests

This also changes package.json to configure separate test scripts for
jest and cypress. It also defines jest.rootDir so jest ignores cypress
test files.

* Add action for frontend Cypress tests

Not totally sure this will work, as it has to set up the frontend and
backend environments.

* Add action for frontend Cypress tests

Not totally sure this will work, as it has to set up the frontend and
backend environments. Also deletes the unused example tests to avoid
running them in the action.

* Add & sign to ensure Go server step doesn't block

* Remove Cypress workflow.

Cypress is unable to access http://localhost:8080 and I'm not sure what
the solution/workaround is. For now, just removing this since Go/Jest
are the most important to automate.