Refactor API to join user/tutor data
This treats Tutor as a proper subtype of User in API responses,
simplifying the handling of relevant routes and avoiding duplicate data
in responses.

The frontend has not been updated for this yet; the PATCH /profile route
and frontend profile page have outstanding PRs that should be merged
first to minimize unnecessary changes.