Refactor API to join user/tutor data (#120)
* Refactor API to join user/tutor data

This treats Tutor as a proper subtype of User in API responses,
simplifying the handling of relevant routes and avoiding duplicate data
in responses.

The frontend has not been updated for this yet; the PATCH /profile route
and frontend profile page have outstanding PRs that should be merged
first to minimize unnecessary changes.

* Fix tutor.User result for /profile

This used the wrong column name and didn't load the User field. Instead
of using a preload/join, this assigns the User directly since it's
already available.

* Update frontend use of user/tutor information

Also includes simplifying App state to just be the user's profile.