Update remaining fetch calls to use Utils.fetchJson for error handling
This also fixes an issue in Utils.fetchJson for responses that do not
have a JSON body. While the backend is entirely JSON, intermediate
errors like the server being unavailable respond with a protocol error
that is raw text. This solution interprets any non-json response with
being an error with the body as the message.