Update Jest tests and change AsyncWrapper state handling
Jest tests needed mocks for more of the response fields used by
Utils.fetchJson. These should eventually be extracted into a helper
function to avoid duplication. Furthermore, tests need to account for
both fetch errors (rejected) and API errors (error response).

Furthermore, there were some issues with infinite reloading of
components due to waiting on nonexistent data because of the new
loading component. This led to moving the data state into AsyncWrapper
and then passing that directly to the child component. It's unclear if
this was fixed or even an issue in the first place, but the state
management of this approach is better anyways.