Add backend username/password validation and cleanup frontend
This does not validate username/password for signin, since it isn't
needed and isn't done on the frontend either (though would be preferable
for more descriptive errors).

For the actual validation, username must start with a letter (standard
requirement for clarity), only supports alphanumeric and _-. characters
(common characters and dividers that are url safe), and no more than 72
character (to match password length). The password has no character
restrictions (restricting or requiring certain characters lowers
security and can be problematic for password managers) and must be 8-72
characters (8 is a reasonable starting point, 72 appears to be the limit
for bcyrpt).