Fix API error handling (#138)
* Proof of concept for loading/error state wrapper component

Instead of each page managing loading/error state, this component takes
an async handler function and component to render if the handler
resolves successfully. Additionally, the Utils.fetchJson method also
incorporates the response status to ensure only ok responses (2xx) are
fully resolved.

Passing the component via children doesn't work because they are eagerly
rendered before being passed to the component and thus would need to
check for data availability manually.

* Apply AsyncWrapper to other pages

* Update remaining fetch calls to use Utils.fetchJson for error handling

This also fixes an issue in Utils.fetchJson for responses that do not
have a JSON body. While the backend is entirely JSON, intermediate
errors like the server being unavailable respond with a protocol error
that is raw text. This solution interprets any non-json response with
being an error with the body as the message.

* Update Jest tests and change AsyncWrapper state handling

Jest tests needed mocks for more of the response fields used by
Utils.fetchJson. These should eventually be extracted into a helper
function to avoid duplication. Furthermore, tests need to account for
both fetch errors (rejected) and API errors (error response).

Furthermore, there were some issues with infinite reloading of
components due to waiting on nonexistent data because of the new
loading component. This led to moving the data state into AsyncWrapper
and then passing that directly to the child component. It's unclear if
this was fixed or even an issue in the first place, but the state
management of this approach is better anyways.