Fix profile changes not refreshing components
Actions that update the profile (signup/signin, adding a course, and
editing the profile) must ensure any corresponding state is refreshed.
This solution uses setProfile(undefined) to force reload the profile and
ensures that it is used as a dependency for any relevant useEffects.

Ideally, we would keep the frontend state updated as changes are applied
for things like CoursePage rather than fully refreshing, but that's an
optimzation for another time.