Implement availability editing in PATCH /profile
Due to the merging behavior on the frontend, including the availability
will update the entire tutor's availability instead of patching
individual entries. Additionally, there is not currently validation on
the values passed in. Times are stored in HH:MM format based on
https://sqlite.org/lang_datefunc.html.

This also fixes the handling of defaults by using pointers rather than
direct values for fields to allow checking for nil, otherwise it becomes
impossible to clear a value from the user's profile.