Implement backend support for modifying availability (#149)
* Proof-of-concept for Tutor.Availability with many-to-one table

Still has a few hacky solutions for json.Marshal and loading via
Prefetch, but does work.

* Implement availability editing in PATCH /profile

Due to the merging behavior on the frontend, including the availability
will update the entire tutor's availability instead of patching
individual entries. Additionally, there is not currently validation on
the values passed in. Times are stored in HH:MM format based on
https://sqlite.org/lang_datefunc.html.

This also fixes the handling of defaults by using pointers rather than
direct values for fields to allow checking for nil, otherwise it becomes
impossible to clear a value from the user's profile.

* Connect frontend and backend availability handling

Changed the backend schema to fully match the frontend - fields are easy
enough, but the time conversion is particularly messy. Long term HH:MM
should be used to support SQLite time functions (hh:MM AM/PM does not).

Tests also were modified to always include availability in the request.
Long term, this should only include the availability if it has been
modified (which goes along with only tracking changes to availability).